+++ ExpandWorldPrefabs/ExpandWorldPrefabs.cs
@@ -7,12 +7,13 @@ using HarmonyLib;
 using Service;
 using UnityEngine;
 namespace ExpandWorld.Prefab;
+
 [BepInPlugin(GUID, NAME, VERSION)]
 public class EWP : BaseUnityPlugin
 {
   public const string GUID = "expand_world_prefabs";
   public const string NAME = "Expand World Prefabs";
-  public const string VERSION = "1.41";
+  public const string VERSION = "1.41.2";
 #nullable disable
   public static Harmony Harmony;
 #nullable enable