+++ publish/CHANGELOG.md
@@ -1,3 +1,8 @@
+- v1.42
+  - Adds a new field `filterLimit` to set how many filters must match.
+  - Fixes comment or extra space at end of the fields `spawn` or `swap` breaking the file parsing.
+  - Reworks the filtering system. Now multiple filters are checked separately, instead of being combined into a single check.
+
 - v1.41
   - Changes the field `poke` and `objects` to use `filter` instead of `data` (old way still works).
   - Fixes multiple values not working for the field `data`.